requirements.txt
torch>=2.6.0
numpy>=1.26.4
pandas>=2.2.1
scikit-learn>=1.4.1
mlflow>=2.11.1
evidently>=0.4.28
great-expectations>=0.18.12
fastapi>=0.110.0
uvicorn>=0.27.1
streamlit>=1.31.1
plotly>=5.18.0
pyyaml>=6.0.1
openai>=1.13.3
requests>=2.32.3
pydantic>=2.6.3
joblib>=1.3.2
httpx>=0.27.2
python-dotenv>=1.0.1
